refactor: MultimediaDto -> List<MultimediaDto>
