+++ src/main/java/com/pyro/yolog/domain/memo/dto/request/MemoRequest.java
@@ -1,9 +1,12 @@
 package com.pyro.yolog.domain.memo.dto.request;
 
+import com.pyro.yolog.domain.memo.dto.MultimediaDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 @Getter
 @AllArgsConstructor
@@ -12,5 +15,5 @@ import lombok.NoArgsConstructor;
 public class MemoRequest {
     private String title;
     private String content;
-    private MultimediaRequest multimediaRequest;
+    private List<MultimediaDto> multimediaDtos;
 }
